Refactor: Message to kafka ProducerRecord conversion logic extracted from KafkaBrokerMessageProducer
