+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/producer/kafka/KafkaBrokerMessageProducer.java
@@ -1,32 +1,20 @@
 package pl.allegro.tech.hermes.frontend.producer.kafka;
 
-import org.apache.avro.Schema;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.PartitionInfo;
-import org.apache.kafka.common.header.Header;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import pl.allegro.tech.hermes.api.Topic;
-import pl.allegro.tech.hermes.common.config.ConfigFactory;
-import pl.allegro.tech.hermes.common.config.Configs;
 import pl.allegro.tech.hermes.common.metric.HermesMetrics;
 import pl.allegro.tech.hermes.frontend.metric.CachedTopic;
 import pl.allegro.tech.hermes.frontend.producer.BrokerMessageProducer;
 import pl.allegro.tech.hermes.frontend.publishing.PublishingCallback;
 import pl.allegro.tech.hermes.frontend.publishing.message.Message;
-import pl.allegro.tech.hermes.schema.CompiledSchema;
-import pl.allegro.tech.hermes.schema.SchemaId;
-import pl.allegro.tech.hermes.schema.SchemaVersion;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static java.util.Optional.ofNullable;
 
 @Singleton
 public class KafkaBrokerMessageProducer implements BrokerMessageProducer {
@@ -35,26 +23,23 @@ public class KafkaBrokerMessageProducer implements BrokerMessageProducer {
     private final Producers producers;
     private final KafkaTopicMetadataFetcher kafkaTopicMetadataFetcher;
     private final HermesMetrics metrics;
-    private final KafkaHeaderFactory kafkaHeaderFactory;
-    private final boolean schemaIdHeaderEnabled;
+    private final MessageToKafkaProducerRecordConverter messageConverter;
 
     @Inject
     public KafkaBrokerMessageProducer(Producers producers,
                                       KafkaTopicMetadataFetcher kafkaTopicMetadataFetcher,
                                       HermesMetrics metrics,
-                                      KafkaHeaderFactory kafkaHeaderFactory,
-                                      ConfigFactory configFactory) {
+                                      MessageToKafkaProducerRecordConverter messageConverter) {
         this.producers = producers;
         this.kafkaTopicMetadataFetcher = kafkaTopicMetadataFetcher;
         this.metrics = metrics;
-        this.kafkaHeaderFactory = kafkaHeaderFactory;
-        this.schemaIdHeaderEnabled = configFactory.getBooleanProperty(Configs.SCHEMA_ID_HEADER_ENABLED);
+        this.messageConverter = messageConverter;
         producers.registerGauges(metrics);
     }
 
     @Override
     public void send(Message message, CachedTopic cachedTopic, final PublishingCallback callback) {
-        ProducerRecord<byte[], byte[]> producerRecord = createProducerRecord(message, cachedTopic);
+        ProducerRecord<byte[], byte[]> producerRecord = messageConverter.convertToProducerRecord(message, cachedTopic.getKafkaTopics().getPrimary().name());
 
         try {
             producers.get(cachedTopic.getTopic()).send(producerRecord, new SendCallback(message, cachedTopic.getTopic(), callback));
@@ -62,38 +47,6 @@ public class KafkaBrokerMessageProducer implements BrokerMessageProducer {
             // message didn't get to internal producer buffer and it will not be send to a broker
             callback.onUnpublished(message, cachedTopic.getTopic(), e);
         }
-    }
-
-    private ProducerRecord<byte[], byte[]> createProducerRecord(Message message, CachedTopic cachedTopic) {
-        String kafkaTopicName = cachedTopic.getKafkaTopics().getPrimary().name().asString();
-        Optional<SchemaVersion> schemaVersion = message.<Schema>getCompiledSchema().map(CompiledSchema::getVersion);
-        Optional<SchemaId> schemaId = createSchemaId(message);
-        Iterable<Header> headers = createRecordHeaders(message.getId(), message.getTimestamp(), schemaId, schemaVersion);
-        byte[] partitionKey = ofNullable(message.getPartitionKey()).map(String::getBytes).orElse(null);
-
-        return new ProducerRecord<byte[], byte[]>(kafkaTopicName, null, partitionKey, message.getData(), headers);
-    }
-
-    private Optional<SchemaId> createSchemaId(Message message) {
-        if (schemaIdHeaderEnabled) {
-            return message.<Schema>getCompiledSchema().map(CompiledSchema::getId);
-        }
-
-        return Optional.empty();
-    }
-
-    private Iterable<Header> createRecordHeaders(String id, long timestamp, Optional<SchemaId> schemaId, Optional<SchemaVersion> schemaVersion) {
-        Stream<Optional<Header>> headers = Stream.of(
-            Optional.of(kafkaHeaderFactory.messageId(id)),
-            Optional.of(kafkaHeaderFactory.timestamp(timestamp)),
-            schemaVersion.map(sv -> kafkaHeaderFactory.schemaVersion(sv.value())),
-            schemaId.map(sid -> kafkaHeaderFactory.schemaId(sid.value()))
-        );
-
-        return headers
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .collect(Collectors.toList());
     }
 
     @Override
