+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/producer/kafka/KafkaBrokerMessageProducerFactory.java
@@ -1,7 +1,6 @@
 package pl.allegro.tech.hermes.frontend.producer.kafka;
 
 import org.glassfish.hk2.api.Factory;
-import pl.allegro.tech.hermes.common.config.ConfigFactory;
 import pl.allegro.tech.hermes.common.metric.HermesMetrics;
 
 import javax.inject.Inject;
@@ -11,25 +10,22 @@ public class KafkaBrokerMessageProducerFactory implements Factory<KafkaBrokerMes
     private final Producers producers;
     private final KafkaTopicMetadataFetcher kafkaTopicMetadataFetcher;
     private final HermesMetrics hermesMetrics;
-    private final KafkaHeaderFactory kafkaHeaderFactory;
-    private final ConfigFactory configFactory;
+    private final MessageToKafkaProducerRecordConverter messageConverter;
 
     @Inject
     public KafkaBrokerMessageProducerFactory(Producers producers,
                                              KafkaTopicMetadataFetcher kafkaTopicMetadataFetcher,
                                              HermesMetrics hermesMetrics,
-                                             KafkaHeaderFactory kafkaHeaderFactory,
-                                             ConfigFactory configFactory) {
+                                             MessageToKafkaProducerRecordConverter messageConverter) {
         this.producers = producers;
         this.kafkaTopicMetadataFetcher = kafkaTopicMetadataFetcher;
         this.hermesMetrics = hermesMetrics;
-        this.kafkaHeaderFactory = kafkaHeaderFactory;
-        this.configFactory = configFactory;
+        this.messageConverter = messageConverter;
     }
 
     @Override
     public KafkaBrokerMessageProducer provide() {
-        return new KafkaBrokerMessageProducer(producers, kafkaTopicMetadataFetcher, hermesMetrics, kafkaHeaderFactory, configFactory);
+        return new KafkaBrokerMessageProducer(producers, kafkaTopicMetadataFetcher, hermesMetrics, messageConverter);
     }
 
     @Override
