+++ hermes-frontend/src/test/java/pl/allegro/tech/hermes/frontend/producer/kafka/KafkaBrokerMessageProducerTest.java
@@ -59,7 +59,8 @@ public class KafkaBrokerMessageProducerTest {
     @Before
     public void before() {
         cachedTopic = new CachedTopic(TOPIC, hermesMetrics, kafkaNamesMapper.toKafkaTopics(TOPIC));
-        producer = new KafkaBrokerMessageProducer(producers, kafkaTopicMetadataFetcher, hermesMetrics, kafkaHeaderFactory, configFactory);
+        MessageToKafkaProducerRecordConverter messageConverter = new MessageToKafkaProducerRecordConverter(kafkaHeaderFactory, configFactory);
+        producer = new KafkaBrokerMessageProducer(producers, kafkaTopicMetadataFetcher, hermesMetrics, messageConverter);
     }
 
     @After
